v0.0.4 - Game phases
Added phases to the game. The players can move on the move phase, and catch pokemon (not yet implemented) on the catch phase. Players get to roll the dice once on the movement phase and can move the rolled number of spaces. All players must click the ready button for the game to go to the next phase