Inventory System
Added inventory system for caught pokemon
Pokemon are auto added to inventory (max 6)
If there is already 6 and the new pokemon is stronger, the weakest is replaced.
Pokemon in inventory change what dice you roll. When you roll the dice, the game picks a random pokemon in your party and rolls their dice.